test(#32): Update PostTransformer
- Clean code
- Get the shortValue from PostStatus attributes

+++ common/src/test/java/org/wcdevs/blog/core/common/post/PostTransformerTest.java
@@ -239,7 +239,7 @@ class PostTransformerTest {
                      .updatedOn(updatedOn)
                      .publishedBy(publishedBy)
                      .updatedBy(updatedBy)
-                     .status(PostStatus.PUBLISHED)
+                     .status(PostStatus.PUBLISHED.shortValue())
                      .build();
 
     var dto = new PostTransformer().dtoFromEntity(entity);
